requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2
django-ckeditor-5==0.2.17
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.16.0
dlib==20.0.0
drf-yasg==1.21.10
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.10
inflection==0.5.1
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
